refactor: move enums around
