+++ services/121-service/src/notifications/enum/process-names.enum.ts
@@ -1,0 +1,6 @@
+// Queues for sending messages
+export enum ProcessNameMessage {
+  send = 'send',
+  sms = 'sms',
+  whatsapp = 'whatsapp',
+}
