+++ services/121-service/src/notifications/message-queues/message-queues.service.ts
@@ -3,12 +3,8 @@ import { InjectRepository } from '@nestjs/typeorm';
 import { Queue } from 'bull';
 import { Equal, Repository } from 'typeorm';
 
-import {
-  DEFAULT_QUEUE_CREATE_MESSAGE,
-  MESSAGE_QUEUE_MAP,
-  MessageQueueMap,
-} from '@121-service/src/notifications/enum/message-queue-mapping.const';
 import { MessageContentType } from '@121-service/src/notifications/enum/message-type.enum';
+import { ProcessNameMessage } from '@121-service/src/notifications/enum/process-names.enum';
 import {
   ExtendedMessageProccessType,
   MessageJobCustomDataDto,
@@ -16,18 +12,20 @@ import {
   MessageProcessType,
   MessageProcessTypeExtension,
 } from '@121-service/src/notifications/message-job.dto';
+import {
+  DEFAULT_QUEUE_CREATE_MESSAGE,
+  MESSAGE_QUEUE_MAP,
+  MessageQueueMap,
+} from '@121-service/src/notifications/message-queue-mapping.const';
 import { MessageTemplateEntity } from '@121-service/src/notifications/message-template/message-template.entity';
 import { ProgramAttributesService } from '@121-service/src/program-attributes/program-attributes.service';
+import { QueueNameCreateMessage } from '@121-service/src/queues/enum/queue-names-create-message.enum';
 import { QueuesService } from '@121-service/src/queues/queues.service';
 import { DefaultRegistrationDataAttributeNames } from '@121-service/src/registration/enum/registration-attribute.enum';
 import { RegistrationDataService } from '@121-service/src/registration/modules/registration-data/registration-data.service';
 import { RegistrationEntity } from '@121-service/src/registration/registration.entity';
 import { RegistrationViewEntity } from '@121-service/src/registration/registration-view.entity';
 import { LanguageEnum } from '@121-service/src/shared/enum/language.enums';
-import {
-  ProcessNameMessage,
-  QueueNameCreateMessage,
-} from '@121-service/src/shared/enum/queue-process.names.enum';
 
 @Injectable()
 export class MessageQueuesService {
