+++ services/121-service/src/notifications/processors/message-incoming.processor.ts
@@ -1,11 +1,9 @@
 import { Process, Processor } from '@nestjs/bull';
 import { Job } from 'bull';
 
+import { ProcessNameMessage } from '@121-service/src/notifications/enum/process-names.enum';
 import { MessageIncomingService } from '@121-service/src/notifications/message-incoming/message-incoming.service';
-import {
-  ProcessNameMessage,
-  QueueNameMessageCallBack,
-} from '@121-service/src/shared/enum/queue-process.names.enum';
+import { QueueNameMessageCallBack } from '@121-service/src/queues/enum/queue-names-message-callback.enum';
 import { AzureLogService } from '@121-service/src/shared/services/azure-log.service';
 @Processor(QueueNameMessageCallBack.incomingMessage)
 export class MessageIncomingProcessor {
