+++ services/121-service/src/queues/enum/queue-names-create-message.enum.ts
@@ -1,0 +1,7 @@
+export enum QueueNameCreateMessage {
+  replyOnIncoming = 'replyOnIncoming',
+  smallBulk = 'smallBulk',
+  mediumBulk = 'mediumBulk',
+  largeBulk = 'largeBulk',
+  lowPriority = 'lowPriority',
+}
