+++ services/121-service/src/queues/enum/queue-names-message-callback.enum.ts
@@ -1,0 +1,4 @@
+export enum QueueNameMessageCallBack {
+  status = 'messageStatusCallback',
+  incomingMessage = 'incomingMessage',
+}
