+++ services/121-service/src/queues/enum/queue-names-registration.enum.ts
@@ -1,0 +1,4 @@
+// Queues for updating registrations
+export enum QueueNameRegistration {
+  registration = 'registration',
+}
