+++ services/121-service/src/queues/enum/queue-names-safaricom-callback.enum.ts
@@ -1,4 +1,4 @@
-export enum SafaricomCallbackQueueNames {
+export enum QueueNamesSafaricomCallback {
   transfer = 'safaricomTransferCallback',
   timeout = 'safaricomTimeoutCallback',
 }
