+++ services/121-service/src/queues/enum/queue-names-transaction-jobs.enum.ts
@@ -1,5 +1,5 @@
 // Refactor once all FSPs are migrated to new architecture: then this can move from 'shared' to transaction-queue module
-export enum TransactionJobQueueNames {
+export enum QueueNamesTransactionJob {
   intersolveVisa = 'intersolveVisa-transaction-jobs',
   intersolveVoucher = 'intersolveVoucher-transaction-jobs',
   commercialBankEthiopia = 'commercialBankEthiopia-transaction-jobs',
