+++ services/121-service/src/registration/enum/process-name-registration.enum.ts
@@ -1,0 +1,3 @@
+export enum ProcessNameRegistration {
+  update = 'updateRegistration',
+}
