+++ services/121-service/src/registration/processsors/registrations-update.processor.ts
@@ -2,13 +2,11 @@ import { Process, Processor } from '@nestjs/bull';
 import { Scope } from '@nestjs/common';
 import { Job } from 'bull';
 
+import { QueueNameRegistration } from '@121-service/src/queues/enum/queue-names-registration.enum';
 import { RegistrationsUpdateJobDto } from '@121-service/src/registration/dto/registration-update-job.dto';
 import { UpdateRegistrationDto } from '@121-service/src/registration/dto/update-registration.dto';
+import { ProcessNameRegistration } from '@121-service/src/registration/enum/process-name-registration.enum';
 import { RegistrationsService } from '@121-service/src/registration/registrations.service';
-import {
-  ProcessNameRegistration,
-  QueueNameRegistration,
-} from '@121-service/src/shared/enum/queue-process.names.enum';
 @Processor({
   name: QueueNameRegistration.registration,
   scope: Scope.REQUEST,
