+++ services/121-service/src/transaction-job-processors/processors/transaction-job-intersolve-visa.processor.ts
@@ -7,11 +7,11 @@ import {
   getRedisSetName,
   REDIS_CLIENT,
 } from '@121-service/src/payments/redis/redis-client';
+import { QueueNamesTransactionJob } from '@121-service/src/queues/enum/queue-names-transaction-jobs.enum';
 import { JobNames } from '@121-service/src/shared/enum/job-names.enum';
-import { TransactionJobQueueNames } from '@121-service/src/shared/enum/transaction-job-queue-names.enum';
 import { TransactionJobProcessorsService } from '@121-service/src/transaction-job-processors/transaction-job-processors.service';
 
-@Processor(TransactionJobQueueNames.intersolveVisa)
+@Processor(QueueNamesTransactionJob.intersolveVisa)
 export class TransactionJobProcessorIntersolveVisa {
   constructor(
     private readonly transactionJobProcessorsService: TransactionJobProcessorsService,
